fix(dashboards): keep custom date ranges beyond 90 days
The filters bar derives the date picker's `maxPickableDays` from the
dashboard's widget data categories, which can exceed 90 days for spans
with downsampled retention. The `PageFiltersContainer` wrapping the
dashboard passed no limit at all, so it fell back to the plan's
`retentionDays` (90) and reset any absolute range starting before that
window back to `90d` — immediately after the picker allowed it.

Feed the container the same widget-derived limit the picker uses.
