+++ static/app/views/dashboards/filtersBar.tsx
@@ -1,4 +1,4 @@
-import {useEffect, useMemo, useState} from 'react';
+import {useEffect, useState} from 'react';
 import styled from '@emotion/styled';
 import type {Location} from 'history';
 import {createParser, useQueryState} from 'nuqs';
@@ -23,11 +23,9 @@ import {
 } from 'sentry/constants/releases';
 import {IconAdd, IconClock} from 'sentry/icons';
 import {t} from 'sentry/locale';
-import {DataCategory} from 'sentry/types/core';
 import type {User} from 'sentry/types/user';
 import {trackAnalytics} from 'sentry/utils/analytics';
 import {defined} from 'sentry/utils/defined';
-import {useMaxPickableDays} from 'sentry/utils/useMaxPickableDays';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {useUser} from 'sentry/utils/useUser';
 import {useUserTeams} from 'sentry/utils/useUserTeams';
@@ -39,6 +37,7 @@ import {
   mergeGlobalFilters,
 } from 'sentry/views/dashboards/globalFilter/utils';
 import {useDashboardChartInterval} from 'sentry/views/dashboards/hooks/useDashboardChartInterval';
+import {useDashboardMaxPickableDays} from 'sentry/views/dashboards/hooks/useDashboardMaxPickableDays';
 import {useDatasetSearchBarData} from 'sentry/views/dashboards/hooks/useDatasetSearchBarData';
 import {useInvalidateStarredDashboards} from 'sentry/views/dashboards/hooks/useInvalidateStarredDashboards';
 import {getDashboardFiltersFromURL} from 'sentry/views/dashboards/utils';
@@ -55,53 +54,8 @@ import type {
   DashboardFilters,
   DashboardPermissions,
   GlobalFilter,
-  Widget,
 } from './types';
-import {DashboardFilterKeys, WidgetType} from './types';
-
-/**
- * Maps widget types to data categories for determining max pickable days
- */
-function getDataCategoriesFromWidgets(
-  widgets: Widget[]
-): [DataCategory, ...DataCategory[]] {
-  const categories = new Set<DataCategory>();
-
-  for (const widget of widgets) {
-    const widgetType = widget.widgetType ?? WidgetType.DISCOVER;
-
-    switch (widgetType) {
-      case WidgetType.SPANS:
-        categories.add(DataCategory.SPANS);
-        break;
-      case WidgetType.TRANSACTIONS:
-        categories.add(DataCategory.TRANSACTIONS);
-        break;
-      case WidgetType.TRACEMETRICS:
-        categories.add(DataCategory.TRACE_METRICS);
-        break;
-      case WidgetType.LOGS:
-        categories.add(DataCategory.LOG_ITEM);
-        break;
-      case WidgetType.ERRORS:
-      case WidgetType.DISCOVER:
-      case WidgetType.ISSUE:
-      case WidgetType.RELEASE:
-      case WidgetType.METRICS:
-      default:
-        // For error-like widgets, use TRANSACTIONS as a safe default
-        // since it has the most permissive date range
-        categories.add(DataCategory.TRANSACTIONS);
-        break;
-    }
-  }
-
-  // Return as tuple with at least one element (required by useMaxPickableDays)
-  const categoriesArray = Array.from(categories);
-  return categoriesArray.length > 0
-    ? (categoriesArray as [DataCategory, ...DataCategory[]])
-    : [DataCategory.TRANSACTIONS];
-}
+import {DashboardFilterKeys} from './types';
 
 export type FiltersBarProps = {
   filters: DashboardFilters;
@@ -149,18 +103,7 @@ export function FiltersBar({
     ? (PREBUILT_DASHBOARDS[prebuiltDashboardId].filters.globalFilter ?? [])
     : [];
 
-  // Determine data categories based on widget types in the dashboard
-  const dataCategories = useMemo(() => {
-    if (!dashboard?.widgets || dashboard.widgets.length === 0) {
-      // Default to TRANSACTIONS if no widgets
-      return [DataCategory.TRANSACTIONS] as [DataCategory, ...DataCategory[]];
-    }
-
-    return getDataCategoriesFromWidgets(dashboard.widgets);
-  }, [dashboard?.widgets]);
-
-  // Calculate maxPickableDays based on the data categories
-  const maxPickableDaysOptions = useMaxPickableDays({dataCategories});
+  const maxPickableDaysOptions = useDashboardMaxPickableDays(dashboard?.widgets);
 
   // Release sort state - validates and defaults to DATE via custom parser
   const [releaseSort, setReleaseSort] = useQueryState('sortReleasesBy', parseReleaseSort);
